requirements.txt
Flask==3.0.0
gunicorn
supabase
python-dotenv